refactor: separate query to utils/query
