+++ lua/nvim-treesitter/query.lua
@@ -1,6 +1,7 @@
 local api = vim.api
+local qutils = require'nvim-treesitter.utils.query'
 local ts = vim.treesitter
--- local locals = require'nvim-treesitter.locals'
+local locals = require'nvim-treesitter.locals'
 
 local M = {}
 
@@ -73,67 +74,7 @@ function M.get_query(lang, query_name)
 end
 
 function M.iter_prepared_matches(query, qnode, bufnr, start_row, end_row)
-  -- A function that splits  a string on '.'
-  local function split(string)
-    local t = {}
-    for str in string.gmatch(string, "([^.]+)") do
-      table.insert(t, str)
-    end
-
-    return t
-  end
-
-  -- Given a path (i.e. a List(String)) this functions inserts value at path
-  local function insert_to_path(object, path, value)
-    local curr_obj = object
-
-    for index=1,(#path -1) do
-      if curr_obj[path[index]] == nil then
-        curr_obj[path[index]] = {}
-      end
-
-      curr_obj = curr_obj[path[index]]
-    end
-
-    curr_obj[path[#path]] = value
-  end
-
-  local matches = query:iter_matches(qnode, bufnr, start_row, end_row)
-
-  local function iter()
-    local pattern, match = matches()
-    if pattern ~= nil then
-      local prepared_match = {}
-
-      -- Extract capture names from each match
-      for id, node in pairs(match) do
-        local name = query.captures[id] -- name of the capture in the query
-        if name ~= nil then
-          local path = split(name)
-          insert_to_path(prepared_match, path, { node=node })
-        end
-      end
-
-      -- Add some predicates for testing
-      local preds = query.info.patterns[pattern]
-      if preds then
-        for _, pred in pairs(preds) do
-          if pred[1] == "set!" and type(pred[2]) == "string" then
-            insert_to_path(prepared_match, split(pred[2]), pred[3])
-          end
-          if pred[1] == "is?" and type(pred[3]) == "string" then
-            if not locals.is(pred[2], pred[3], bufnr) then
-              return iter() -- We should ignore this one, tail call
-            end
-          end
-        end
-      end
-
-      return prepared_match
-    end
-  end
-
-  return iter
+  return qutils.iter_prepared_matches(match_pred, query, qnode, bufnr, start_row, end_row)
 end
 
 return M
